# docs/api/openapi/alloy-public-api.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Alloy Public API",
    "version": "1.0.0",
    "x-spec-status": "public-v1",
    "description": "The Alloy Public API. This artifact describes ONLY endpoints that exist. Coverage is a build gate, not a report: a documented path with no route, or a route under /api/v1 with no documentation, fails the drift guard in web/tests/platform/external/openApiDriftGuard.test.ts.\n\nAuthority is never chosen by the caller. An access token resolves to exactly one installation, and that installation determines the organization, the granted scopes and the resource boundary. There is no org_id parameter anywhere in this contract."
  },
  "servers": [
    {
      "url": "https://{host}",
      "variables": {
        "host": {
          "default": "app.alloy.example"
        }
      }
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/schedule-assignments": {
      "get": {
        "operationId": "listScheduleAssignments",
        "tags": [
          "Service state"
        ],
        "summary": "List committed schedule assignments",
        "description": "The committed, effective-dated schedule commitment for a child. Requires the `schedule.read` scope.\n\n**This is the canonical schedule resource.** It is persisted, it is synchronizable, and it tells you what was actually committed: which recurring pattern applies to a child, at which site and room, over which dates.\n\n**Choosing between the two schedule endpoints:**\n\n- Use **this** endpoint when you hold standing intent and need to be told when it changes. It participates fully in incremental synchronization.\n- Use **`GET /api/v1/schedule-days`** when you are asking \"who is expected on Tuesday\". That endpoint derives individual days from these assignments and is a windowed query, not a collection.\n\nOnly children appear here. Staff schedules are not part of this resource.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-required-scope": "schedule.read",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            },
            "description": "Page size. Clamped to 200."
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Opaque cursor from a previous `next_cursor`. A cursor is a position, never a permission."
          },
          {
            "name": "since_token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Resume exactly where a previous pass stopped. Pass back the `sync_token` from the last page you consumed. Unlike `updated_since` it identifies a single row, so two records sharing a timestamp cannot be skipped or repeated. Prefer this for incremental synchronization."
          },
          {
            "name": "updated_since",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Return only records changed strictly after this instant. Must carry a timezone offset or Z. Sub-second precision is honoured exactly as supplied. `since_token` is preferred: a timestamp alone cannot separate two records that share one."
          },
          {
            "name": "child_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Return only assignments for this child."
          },
          {
            "name": "site_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Return only assignments at this site."
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Return only assignments in this state."
          }
        ],
        "responses": {
          "200": {
            "description": "A page of results.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleAssignmentPage"
                }
              }
            }
          },
          "401": {
            "description": "The access token is missing, expired or not valid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The installation does not hold the required scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/schedule-days": {
      "get": {
        "operationId": "listScheduleDays",
        "tags": [
          "Service state"
        ],
        "summary": "Project expected days from committed schedules",
        "description": "Answers \"who is expected on Tuesday\". Requires the `schedule.read` scope.\n\n**These rows are derived, not stored.** Each one is generated from a committed schedule assignment and its recurring pattern. They have no identifier of their own and no change history, which is why this endpoint deliberately returns **no cursor and no sync token**. Do not attempt to synchronize it incrementally \u2014 there is nothing to checkpoint against, and any watermark would be one this API could not honour.\n\nIf you need change detection, synchronize `GET /api/v1/schedule-assignments` instead and re-project the window yourself. Each returned day names the assignment it came from in `schedule_assignment_id` so you can do exactly that.\n\n**Deterministic.** The same window over the same committed schedules always returns the same days in the same order.\n\n`from` and `to` are required and the window may not exceed 92 days.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-required-scope": "schedule.read",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "First day of the window, inclusive."
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Last day of the window, inclusive. At most 92 days after `from`."
          },
          {
            "name": "child_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Project only this child's days."
          },
          {
            "name": "site_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Project only days at this site."
          }
        ],
        "responses": {
          "200": {
            "description": "The projected window.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleDayWindow"
                }
              }
            }
          },
          "400": {
            "description": "The window is missing, inverted or longer than 92 days.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "The access token is missing, expired or not valid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The installation does not hold the required scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "An opaque access token from /api/v1/oauth/token. Not a JWT; do not parse it."
      }
    },
...
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "type",
              "message",
              "request_id"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "Stable and machine-readable. Branch on this."
              },
              "type": {
                "type": "string",
                "enum": [
                  "invalid_request",
                  "unauthenticated",
                  "forbidden_scope",
                  "forbidden_resource",
                  "not_found",
                  "conflict",
                  "rate_limited",
                  "internal_error"
                ]
              },
              "message": {
                "type": "string",
                "description": "For humans. NOT a contract; it may change."
              },
              "request_id": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      },
...
      "ScheduleAssignment": {
        "type": "object",
        "description": "The committed, effective-dated schedule commitment for a child \u2014 standing intent, not individual days. Only children appear here.",
        "properties": {
          "id": {
            "type": "string"
          },
          "enrollment_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "child_id": {
            "type": "string"
          },
          "site_location_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "room_location_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "schedule_pattern_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "pattern_label": {
            "type": [
              "string",
              "null"
            ],
            "description": "The operator's name for the recurring pattern, such as a full-week or part-week schedule."
          },
          "schedule_type_key": {
            "type": [
              "string",
              "null"
            ]
          },
          "weekdays": {
            "type": "array",
            "items": {
              "type": "integer",
              "minimum": 0,
              "maximum": 6
            },
            "description": "Days the pattern recurs on, 0 = Sunday."
          },
          "status": {
            "type": [
              "string",
              "null"
            ]
          },
          "commitment_kind": {
            "type": [
              "string",
              "null"
            ]
          },
          "is_primary": {
            "type": "boolean"
          },
          "start_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date"
          },
          "end_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date"
          },
          "supersedes_assignment_id": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "child_id",
          "weekdays",
          "is_primary"
        ]
      },
      "ScheduleAssignmentPage": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleAssignment"
            }
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Pass back as `cursor` for the next page. Null on the last page."
          },
          "sync_token": {
            "type": [
              "string",
              "null"
            ],
            "description": "Checkpoint of the last row on this page. Store it and pass it back as `since_token` to resume exactly here."
          }
        },
        "required": [
          "data",
          "next_cursor",
          "sync_token"
        ]
      },
      "ScheduleDay": {
        "type": "object",
        "description": "One expected day for one child, generated from a committed schedule assignment. Not a stored record \u2014 it has no identifier of its own and no change history.",
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "weekday": {
            "type": "integer",
            "minimum": 0,
            "maximum": 6,
            "description": "0 = Sunday."
          },
          "child_id": {
            "type": "string"
          },
          "site_location_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "room_location_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "schedule_assignment_id": {
            "type": "string",
            "description": "The committed assignment this day was derived from. Synchronize that resource to be told when it changes."
          },
          "schedule_pattern_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "schedule_type_key": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "date",
          "weekday",
          "child_id",
          "schedule_assignment_id"
        ]
      },
      "ScheduleDayWindow": {
        "type": "object",
        "description": "A derived window. Deliberately carries no cursor and no sync token.",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleDay"
            }
          },
          "from": {
            "type": "string",
            "format": "date"
          },
          "to": {
            "type": "string",
            "format": "date"
          },
          "derived_from": {
            "type": "string",
            "description": "Names the canonical authority these rows were computed from."
          }
        },
        "required": [
          "data",
          "from",
          "to",
          "derived_from"
        ]
      },
...
    },
...
  }
}